Data Plus Movie
1. Tableau Data Plus Movies Data Challenge
This repository is dedicated to the Tableau Data Plus Movies data challenge. The goal is to analyze and visualize data related to movies, with a focus on exploring connections between various film industry professionals.

2. Data Processing using Python Script
The provided data has been further processed using the Python script included in this repository. The script plays a crucial role in extracting valuable insights from the dataset. To reproduce the results or understand the processing steps, refer to the uploaded Python script.

3. Kevin Bacon Network Analysis
The heart of this project revolves around identifying and analyzing the professional network of Kevin Bacon in the film industry. The Python script meticulously sifts through the dataset to compile a list of individuals who have collaborated with Kevin Bacon. Subsequently, it explores the films these individuals have been a part of, extending to the second degree of connections.

This network analysis provides a comprehensive view of Kevin Bacon's influence and connections within the film industry. The results can be visualized using Tableau for a more intuitive understanding of the relationships among various professionals in the movie business.


https://www.tableau.com/community/movies/starter-kit